Make more elixir-y
I tried to structure this like the book, but
couldn’t get it exactly the same since elixir doesn’t
allow recursive anonymous functions. You could get
round this by passing in the function as in[1], but I think
that’s harder to read.

Instead, use pattern matching on number of arguments which is
a more natural thing to do in elixir.

[1] https://www.reddit.com/r/elixir/comments/41bva5/recursive_anonymous_functions_in_elixir/cz1laa1/