Typed arith
Tried to separate the rules so they appear in the same
order as the book.

I made the catch-all case raise a `NotWellTyped` error since
if it reaches here I think there are syntax problems rather than
type ones. Note that Elixir would raise a `FunctionClauseError`
anyway without this here.