Examples that trigger `ArrayStoreException`.
I wasn't sure what dangers java prevents in the first example by raising
the exception. I think the example isn't too helpful as the new B array
is only used in one place- where it is assigned to `a`.

My second example assigns the B array to another variable to show a
possible danger- accessing bArray after the exception was raised would
be problemantic since it would contain elements that wearn't of type B
or a subtype of type B.